beginners_lab/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
python-dotenv
